separate additional controls into TableHeader
